Fix Get Latest MinIO Image on Tenant Create
