+++ operatorapi/operator_tenants.go
@@ -1040,77 +1040,6 @@ func setImageRegistry(ctx context.Context, req *models.ImageRegistry, clientset 
 	return pullSecretName, nil
 }
 
-// updateTenantAction does an update on the minioTenant by patching the desired changes
-func updateTenantAction(ctx context.Context, operatorClient OperatorClientI, clientset v1.CoreV1Interface, httpCl utils2.HTTPClientI, namespace string, params operator_api.UpdateTenantParams) error {
-	imageToUpdate := params.Body.Image
-	imageRegistryReq := params.Body.ImageRegistry
-
-	minInst, err := operatorClient.TenantGet(ctx, namespace, params.Tenant, metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	// we can take either the `image_pull_secret` of the `image_registry` but not both
-	if params.Body.ImagePullSecret != "" {
-		minInst.Spec.ImagePullSecret.Name = params.Body.ImagePullSecret
-	} else {
-		// update the image pull secret content
-		if _, err := setImageRegistry(ctx, imageRegistryReq, clientset, namespace, params.Tenant); err != nil {
-			restapi.LogError("error setting image registry secret: %v", err)
-			return err
-		}
-	}
-
-	// if image to update is empty we'll use the latest image by default
-	if strings.TrimSpace(imageToUpdate) != "" {
-		minInst.Spec.Image = imageToUpdate
-	} else {
-		im, err := cluster.GetLatestMinioImage(httpCl)
-		// if we can't get the MinIO image, we won' auto-update it unless it's explicit by name
-		if err == nil {
-			minInst.Spec.Image = *im
-		}
-	}
-
-	// Prometheus Annotations
-	currentAnnotations := minInst.Annotations
-	prometheusAnnotations := map[string]string{
-		prometheusPath:   "/minio/prometheus/metrics",
-		prometheusPort:   fmt.Sprint(miniov2.MinIOPort),
-		prometheusScrape: "true",
-	}
-	if params.Body.EnablePrometheus && currentAnnotations != nil {
-		// add prometheus annotations to the tenant
-		minInst.Annotations = addAnnotations(currentAnnotations, prometheusAnnotations)
-		// add prometheus annotations to the each pool
-		if minInst.Spec.Pools != nil {
-			for _, pool := range minInst.Spec.Pools {
-				poolAnnotations := pool.VolumeClaimTemplate.GetObjectMeta().GetAnnotations()
-				pool.VolumeClaimTemplate.GetObjectMeta().SetAnnotations(addAnnotations(poolAnnotations, prometheusAnnotations))
-			}
-		}
-	} else {
-		// remove prometheus annotations to the tenant
-		minInst.Annotations = removeAnnotations(currentAnnotations, prometheusAnnotations)
-		// add prometheus annotations from each pool
-		if minInst.Spec.Pools != nil {
-			for _, pool := range minInst.Spec.Pools {
-				poolAnnotations := pool.VolumeClaimTemplate.GetObjectMeta().GetAnnotations()
-				pool.VolumeClaimTemplate.GetObjectMeta().SetAnnotations(removeAnnotations(poolAnnotations, prometheusAnnotations))
-			}
-		}
-	}
-
-	payloadBytes, err := json.Marshal(minInst)
-	if err != nil {
-		return err
-	}
-	_, err = operatorClient.TenantPatch(ctx, namespace, minInst.Name, types.MergePatchType, payloadBytes, metav1.PatchOptions{})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 // addAnnotations will merge two annotation maps
 func addAnnotations(annotationsOne, annotationsTwo map[string]string) map[string]string {
 	if annotationsOne == nil {
